requirements.txt
asgiref==3.6.0
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.7
django-cors-headers==3.14.0
django-filter==22.1
djangorestframework==3.14.0
docutils==0.19
idna==3.4
importlib-metadata==6.0.0
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.2
pycodestyle==2.10.0
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.14
zipp==3.14.0
